+++ ORStools/proc/directions_points_layer_proc.py
@@ -37,7 +37,7 @@ from qgis.core import (QgsWkbTypes,
                        QgsPointXY,
                        )
 
-from ORStools.common import directions_core, PROFILES, PREFERENCES
+from ORStools.common import directions_core, PROFILES, PREFERENCES, OPTIMIZATION_MODES
 from ORStools.utils import transform, exceptions, logger
 from .base_processing_algorithm import ORSBaseProcessingAlgorithm
 from ..utils.processing import get_params_optimize
@@ -90,10 +90,12 @@ class ORSDirectionsPointsLayerAlgo(ORSBaseProcessingAlgorithm):
                 PREFERENCES,
                 defaultValue=PREFERENCES[0]
             ),
-            QgsProcessingParameterBoolean(
-                name=self.IN_OPTIMIZE,
-                description="Optimize waypoint order (except first and last)",
-                defaultValue=False
+            QgsProcessingParameterEnum(
+                self.IN_OPTIMIZE,
+                "Traveling Salesman",
+                OPTIMIZATION_MODES,
+                defaultValue=None,
+                optional=True,
             )
         ]
 
@@ -104,7 +106,7 @@ class ORSDirectionsPointsLayerAlgo(ORSBaseProcessingAlgorithm):
 
         preference = dict(enumerate(PREFERENCES))[parameters[self.IN_PREFERENCE]]
 
-        optimize = parameters[self.IN_OPTIMIZE]
+        optimization_mode = parameters[self.IN_OPTIMIZE]
 
         # Get parameter values
         source = self.parameterAsSource(
@@ -162,8 +164,8 @@ class ORSDirectionsPointsLayerAlgo(ORSBaseProcessingAlgorithm):
                 break
 
             try:
-                if optimize:
-                    params = get_params_optimize(points, profile)
+                if optimization_mode is not None:
+                    params = get_params_optimize(points, profile, optimization_mode)
                     response = ors_client.request('/optimization', {}, post_json=params)
 
                     sink.addFeature(directions_core.get_output_features_optimization(