+++ ORStools/utils/processing.py
@@ -32,34 +32,47 @@ from qgis.core import QgsPointXY
 from typing import List
 
 from ORStools import BASE_DIR
+from ORStools.common import OPTIMIZATION_MODES
 
 
-def get_params_optimize(point_list: List[QgsPointXY], ors_profile: str) -> dict:
+def get_params_optimize(point_list: List[QgsPointXY], ors_profile: str, mode: int) -> dict:
     """
     Build parameters for optimization endpoint
 
     :param point_list: individual polyline points
     :param ors_profile: ors transport profile to be used
+    :param mode: optimization mode
     """
+    start = end = None
 
-    start = point_list.pop(0)
-    end = point_list.pop(-1)
+    if mode == OPTIMIZATION_MODES.index("Fix End Point"):
+        end = point_list.pop(-1)
+    elif mode == OPTIMIZATION_MODES.index("Fix Start Point"):
+        start = point_list.pop(0)
+    elif mode == OPTIMIZATION_MODES.index("Fix Start and End Point"):
+        start = point_list.pop(0)
+        end = point_list.pop(-1)
+    elif mode == OPTIMIZATION_MODES.index("Round Trip"):
+        start = end = point_list.pop(0)
+
+    vehicle = {
+        "id": 0,
+        "profile": ors_profile
+    }
+
+    if start:
+        vehicle.update({"start": [round(start.x(), 6), round(start.y(), 6)]})
+    if end:
+        vehicle.update({"end": [round(end.x(), 6), round(end.y(), 6)]})
 
     params = {
-        'jobs': list(),
-        'vehicles': [{
-            "id": 0,
-            "profile": ors_profile,
-            "start": [start.x(), start.y()],
-            "end": [end.x(), end.y()]
-        }],
+        'jobs': [{
+            "location": [round(point.x(), 6), round(point.y(), 6)],
+            "id": point_list.index(point)
+        } for point in point_list],
+        'vehicles': [vehicle],
         'options': {'g': True}
     }
-    for point in point_list:
-        params['jobs'].append({
-            "location": [point.x(), point.y()],
-            "id": point_list.index(point)
-        })
 
     return params